move headers into include and fix CMake configuration
