+++ test/CMakeLists.txt
@@ -6,9 +6,6 @@ include("Utils")
 
 add_hFlux_test(hermite_test
   SRCS
-    ../src/FiniteDifferenceWeights.hpp
-    AnalyticField.hpp
-    common.hpp
     hermite_test.cpp
   INPUT
     hermite_test.input
